requirements.txt
click==7.1.2
Flask==1.1.2
greenlet==1.0.0
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.20.2
pandas==1.2.3
protobuf==3.15.6
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.5
threadpoolctl==2.1.0
Werkzeug==1.0.1
